Add files via upload
This is a script to engage in learning curves that balance by a confounder as you increase the number of subjects. In the original paper, we show how models that use machine learning with genetic variants to predict opioid use disorder are confounded across ancestry. We do this by start in a completely confounded case, and balancing at each iteration of the learning curve as we increase the N we are training on.  This method is used in this preprint: 
https://www.medrxiv.org/content/10.1101/2020.09.12.20193342v2